Fixed DUPLEX implementation (#56)
Change suggested by skylogic004 

"Selecting the 2nd set of most distant points was wrong. D is a symmetric matrix, but the code was removing the columns ONLY (corresponding to the first two most distant points). But it should be removing the columns AND rows corresponding to the first two most distant points."